args: Fix u32 -> usize to use as indices

+++ src/g_args.rs
@@ -25,7 +25,7 @@ pub struct GlepArgs {
     pub filename: String,
     pub is_stdin: bool,
     pub pattern: String,
-    pub lines: Vec<u32>,
+    pub lines: Vec<usize>,
 }
 
 pub fn get_args(cli_args: &Vec<String>) -> GlepArgs {
@@ -37,7 +37,7 @@ pub fn get_args(cli_args: &Vec<String>) -> GlepArgs {
     let mut extra_lines = Vec::new();
 
     cli_args[3..].iter().for_each(|line| {
-        extra_lines.push(line.parse::<u32>().unwrap());
+        extra_lines.push(line.parse::<usize>().unwrap());
         // FIXME: Add better error handling
     });
